+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 1.2.1.5
 - NTsync toggle now works with cachyos-proton
 - XL_PATH can be used as an alias for XL_USERDIR
-- Added a workaround for AppImage builds. Umu-launcher causes a memory overflow and cpu usage spike which can hang the system.
+- Added a warning for AppImage builds. Umu-launcher can crash/hang with nvidia card.
 
 ### Sun August 24 2025 Rankyn Bass <rankyn@proton.me>
 1.2.1.4